Enhances timeline span visualization
Improves the visualization of spans in the timeline by introducing color-coded borders based on the span ID.
This allows for easier differentiation and identification of related spans, especially in deeply nested traces.
It also adds a cursor pointer to the "Load More" icon to improve discoverability of the "Load more traces" functionality.