Adds error icon to span
Adds an error icon next to the span child count if any of the span's events contains the word "error".

This helps to quickly identify spans that have errors.

Fix: Fromat

Update to the new structure of event

Displays error icon on span

Calculates whether a span represents an error based on its attributes, specifically looking for `status.code` with a value of `Error`. If found, renders an error icon next to the span. Also adds an e2e test to verify error icon visibility.

Fixed formatting

Removed console.logs